requirements.txt
langchain
langchain-astradb
langchain-huggingface
langchain-groq
langchain-community
langgraph
datasets
pypdf
python-dotenv
flask
prometheus_client